// src/app/Models/MockObjectStore/app.mock.pensioners.ts
/**
 * Created by Amila on 7/16/2017.
 */
export default class MockPensioners {
  NIC: string;
  pensionNumber: string;
  name: string;
  state: string;
  title: string;
  nameInNic: string;
  dob: any;
  nic: string;
  gender: string;
  district: number;
  dsd: number;
  gnd: number;
  addressOne: string;
  addressTwo: string;
  addressThree: string;
  residence: string;
  mobile: string;
  ministry: string;
  service: string;
  grade: string;
  robf: string;
  designation: number;
  cinp: number;
  retirementReason: string;
  dona: string;
}